requirements.txt
#app section aur about section apne hisab se change karliyo ui dekh lena 
# kuch help lagi to call kar diyo
#terminal me pip install -r req.txt run karliyo
sqlalchemy

bcrypt
streamlit
tensorflow
opencv-python
numpy
joblib
pillow
google.generativeai
python-dotenv
datetime